% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carport_utilities.R
\name{cp_read_events}
\alias{cp_read_events}
\title{Reads carport events data into R environment}
\usage{
cp_read_events(file)
}
\arguments{
\item{file}{Path to the local .xlsx file.}
}
\description{
reads carport event data from local .xlsx file into R environment.
}
\details{
This function loads carport event data from an .xlsx file into the R
environment. Please see the documentation of other function in the
\emph{carport utility functions} group for reading load cell data.
}
\seealso{
Other carport utility functions: 
\code{\link{cp_load_loadcells}()},
\code{\link{cp_pull_loadcells}()},
\code{\link{cp_tidy_loadcells}()}
}
\concept{carport utility functions}
