% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carport_utilities.R
\name{cp_tidy_loadcells}
\alias{cp_tidy_loadcells}
\title{Cleans local load cell data files}
\usage{
cp_tidy_loadcells(load_from)
}
\arguments{
\item{load_from}{Path to the local folder where the data is stored.}
}
\value{
.RData file in specified folder containing all data in tidy format.
}
\description{
cleans load cell data files available in local folder.
}
\details{
This function loads carport load cell data files available in a specified
local folder, cleans the data, converts them into tidy format and saves an
uncompressed .rds file containing all observations. Note, that this function
does not import any data into your R environment, please use other function
in the \emph{carport utility functions} group to import the data into your R
environment.
}
\seealso{
Other carport utility functions: 
\code{\link{cp_load_loadcells}()},
\code{\link{cp_pull_loadcells}()},
\code{\link{cp_read_events}()}
}
\concept{carport utility functions}
