% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{key_valid}
\alias{key_valid}
\title{Check if a provided SSH key is valid}
\usage{
key_valid(ssh_key, url, username)
}
\arguments{
\item{ssh_key}{Path to private SSH key for authentication (.ppk file).}

\item{url}{a string giving the URL}

\item{username}{a string giving the user name for authentication (equals the
folder to be read out on the FTP server).}
}
\value{
TRUE if authentication successful; FALSE if not.
}
\description{
checks if a provided SSH key is a valid for authentication on the SFTP server.
}
\details{
This helper function checks if a provided private SSH key is a valid for
authentication on the SFTP server by trying to read out the "user" folder
file list.
}
\seealso{
Other helper functions: 
\code{\link{is_error}()}
}
\concept{helper functions}
