% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carport_utilities.R
\name{cp_pull_loadcells}
\alias{cp_pull_loadcells}
\title{Download carport load cell data}
\usage{
cp_pull_loadcells(save_to, ssh_key)
}
\arguments{
\item{save_to}{Path to the folder where the data should be saved (also used
to check existing vs. new data on the server).}

\item{ssh_key}{Path to private SSH key for authentication (.ppk file).}
}
\value{
Data files in CSV format in the specified folder and a summary message.
}
\description{
copies carport load cell data files as is from UFZ SFTP server to local directory.
}
\details{
This function downloads all (or only new) available load cell data files from
the Revolution Pi running at the UBZ carport green roof. To run the function,
you need a valid private SSH key to authenticate yourself on the FTP server.
The data files are saved in a local directory. Note, that this function does
not import any data into your R environment, please use other function in the
\emph{carport utility functions} group to import the data into your R
environment.
}
\seealso{
Other carport utility functions: 
\code{\link{cp_tidy_loadcells}()}
}
\concept{carport utility functions}
